wip: added pretraining loss function
