+++ src/transformers/loss/loss_utils.py
@@ -23,6 +23,7 @@ from .loss_for_object_detection import ForObjectDetectionLoss, ForSegmentationLo
 from .loss_grounding_dino import GroundingDinoForObjectDetectionLoss
 from .loss_lw_detr import LwDetrForObjectDetectionLoss
 from .loss_rt_detr import RTDetrForObjectDetectionLoss
+from .loss_vit_nepa import ViTNepaPreTrainingLoss
 
 
 def fixed_cross_entropy(
@@ -165,4 +166,5 @@ LOSS_MAPPING = {
     "DFineForObjectDetection": DFineForObjectDetectionLoss,
     "CsmForConditionalGeneration": ForCausalLMLoss,
     "LwDetrForObjectDetection": LwDetrForObjectDetectionLoss,
+    "ViTNepaForPreTraining": ViTNepaPreTrainingLoss,
 }
