+++ src/transformers/models/vit_nepa/modular_vit_nepa.py
@@ -420,8 +420,7 @@ class ViTNepaOutput(ViTOutput):
 class ViTNepaLayer(ViTLayer):
     def __init__(self, config: ViTNepaConfig, drop_path_rate: float = 0.0):
         super().__init__(config)
-        drop_path_prob = config.drop_path_prob
-        self.drop_path = ViTNepaDropPath(drop_path_prob) if drop_path_prob > 0 else nn.Identity()
+        self.drop_path = ViTNepaDropPath(drop_path_rate) if drop_path_rate > 0 else nn.Identity()
         self.layer_scale = ViTNepaLayerScale(config)
 
     def forward(self, hidden_states: torch.Tensor, position_embeddings: torch.Tensor) -> torch.Tensor:
@@ -512,40 +511,6 @@ class ViTNepaModel(ViTModel):
         return BaseModelOutputWithEmbedding(last_hidden_state=sequence_output, input_embedding=embedding_clean)
 
 
-def prediction_loss(h_in, h_out, shift: bool = True):
-    """
-    similarity loss between two hidden states.
-
-    Args:
-        h_in:  [B, T, D]  input hidden states
-        h_out: [B, T, D]  output hidden states (prediction)
-        shift: if True, compare h_out[:, :-1] with h_in[:, 1:]
-               else, compare h_out with h_in (position-wise)
-
-    Returns:
-        scalar loss (negative cosine similarity)
-    """
-    # detach target
-    h_in = h_in.detach()
-
-    if shift:
-        # shift one step forward
-        p = h_out[:, :-1, :]  # predict next
-        z = h_in[:, 1:, :]  # target is next hidden state
-    else:
-        # same-position matching
-        p = h_out
-        z = h_in
-
-    # normalize
-    p = F.normalize(p, dim=-1)
-    z = F.normalize(z, dim=-1)
-
-    # negative cosine similarity
-    loss = -(p * z).sum(dim=-1).mean()
-    return loss
-
-
 class ViTNepaForPreTraining(ViTNepaPreTrainedModel):
     def __init__(self, config: ViTNepaConfig):
         super().__init__(config)
@@ -575,20 +540,15 @@ class ViTNepaForPreTraining(ViTNepaPreTrainedModel):
             head_mask=head_mask,
             output_attentions=output_attentions,
             interpolate_pos_encoding=interpolate_pos_encoding,
-            is_pretraining=True,
             **kwargs,
         )
 
         sequence_input = outputs.input_embedding
         sequence_output = outputs.last_hidden_state
 
-        embedded_loss = prediction_loss(sequence_input, sequence_output)
+        loss = self.loss_function(sequence_input, sequence_output)
 
-        return EmbeddedModelingOutput(
-            loss=embedded_loss,
-            hidden_states=outputs.hidden_states,
-            attentions=outputs.attentions,
-        )
+        return EmbeddedModelingOutput(loss=loss)
 
 
 class ViTNepaForImageClassification(ViTForImageClassification):
